Create README.md
My primary goal is to apply my technical expertise all throughout the full software life cycle to ensure production and services that meet client specifications. During the learning My experience as junior software developer enhanced my abilities in designing, implementing, testing, and upgrading software. One of my objectives is to keep updated with the latest IT trends and technologies. I am confident that if given the opportunity, I can be a useful talent to the company.